+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Calendar Versioning](https://calver.org).
 - [CSE] Added functionality to disable or enable removing registrations when stopping the CSE. This can be configured via the new configuration setting *cse.registration.unregisterWhenStopping*. By default, registrations are removed when stopping the CSE (the old behavior).
 - [CSE] Added resource type definition file to define the resource types and their attributes in an external JSON file. 
 - [CSE] Increased performance when registering a large number of &lt;AE> resources by adding a database table for registered originators.
+- [CSE] Added validation of received response, e.g. in transfer rerequests. 
 - [WEB] Added support for external root path for the web UI. This allows to run the web UI behind a reverse proxy under a specific path or a Kubernetes ingress path. Thanks to Mudassar Khan (xFlow) for the idea. This is an ESTIMED project contribution.
 
 
@@ -29,6 +30,9 @@ and this project adheres to [Calendar Versioning](https://calver.org).
 - [CSE] For location handling the *content* attributes of &lt;contentInstance> resources support now now integer or JSON objects, not strings anymore. The work-around using a string representation was necessary because the *content* attribute was only supporting string values before.
 - [CSE] Internal restructuring how resource types are configured. The resource types are now defined by an external JSON file that is loaded during CSE initialization.
 - [CSE] Changed the internal handling of resource instantiation and optimized the resource factory.
+- [CSE] Certain ID-related attributes are now converted to SP-relative or Absolute format when responding via Mcc and Mcc'. 
+- [CSE] Improved detection of illegal trailing slashes in "to" identifiers. Thanks to Aaron Wiwior for reporting this issue.
+- [CSE] Fixed crash when forwarding a http DISCOVERY request to another CSE. Thanks to Aaron Wiwior for reporting this issue. 
 
 ### Fixed
 - [CSE] Fixed missing validation of mandatory list attributes (1L) when updating resources.